Refactor Bio-Formats code
Trying to split it up into more manageable pieces...
