+++ qupath-extension-bioformats/src/main/java/qupath/lib/images/servers/bioformats/BioFormatsImageServer.java
@@ -4,7 +4,7 @@
  * %%
  * Copyright (C) 2014 - 2016 The Queen's University of Belfast, Northern Ireland
  * Contact: IP Management (ipmanagement@qub.ac.uk)
- * Copyright (C) 2018 - 2024 QuPath developers, The University of Edinburgh
+ * Copyright (C) 2018 - 2026 QuPath developers, The University of Edinburgh
  * %%
  * QuPath is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as
@@ -23,24 +23,13 @@
 
 package qupath.lib.images.servers.bioformats;
 
-import java.nio.channels.ClosedChannelException;
 import loci.common.services.DependencyException;
 import loci.common.services.ServiceException;
-import loci.formats.ClassList;
-import loci.formats.DimensionSwapper;
 import loci.formats.FormatException;
 import loci.formats.FormatTools;
 import loci.formats.IFormatReader;
 import loci.formats.ImageReader;
-import loci.formats.Memoizer;
-import loci.formats.MetadataTools;
-import loci.formats.ReaderWrapper;
-import loci.formats.gui.AWTImageTools;
-import loci.formats.in.DynamicMetadataOptions;
-import loci.formats.in.MetadataOptions;
 import loci.formats.in.ZarrReader;
-import loci.formats.meta.DummyMetadata;
-import loci.formats.meta.MetadataStore;
 import loci.formats.ome.OMEPyramidStore;
 import loci.formats.ome.OMEXMLMetadata;
 import ome.units.UNITS;
@@ -49,9 +38,6 @@ import ome.xml.meta.OMEXMLMetadataRoot;
 import ome.xml.model.Shape;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import picocli.CommandLine;
-import picocli.CommandLine.Option;
-import picocli.CommandLine.Unmatched;
 import qupath.lib.color.ColorModelFactory;
 import qupath.lib.common.ColorTools;
 import qupath.lib.common.GeneralTools;
@@ -77,19 +63,12 @@ import java.awt.image.BufferedImage;
 import java.awt.image.ColorModel;
 import java.io.File;
 import java.io.IOException;
-import java.lang.ref.Cleaner;
-import java.lang.ref.Cleaner.Cleanable;
 import java.lang.reflect.InvocationTargetException;
 import java.net.URI;
 import java.net.URISyntaxException;
-import java.nio.ByteOrder;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -97,73 +76,54 @@ import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
-import java.util.concurrent.ArrayBlockingQueue;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ForkJoinPool;
-import java.util.concurrent.ForkJoinTask;
 import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.IntStream;
 
 /**
  * QuPath ImageServer that uses the Bio-Formats library to read image data.
  * <p>
- * See http://www.openmicroscopy.org/site/products/bio-formats
+ * See <a href="https://www.openmicroscopy.org/bio-formats/">Bio-Formats homepage</a>.
  * <p>
- * See also https://docs.openmicroscopy.org/bio-formats/6.5.1/developers/matlab-dev.html#improving-reading-performance
- * 
+ * See also <a href="https://bio-formats.readthedocs.io/en/stable/developers/matlab-dev.html#improving-reading-performance">Improving reading performance.</a>
+ *
  * @author Pete Bankhead
  *
  */
 public class BioFormatsImageServer extends AbstractTileableImageServer implements PathObjectReader {
 	
 	private static final Logger logger = LoggerFactory.getLogger(BioFormatsImageServer.class);
-		
-	/**
-	 * Define a maximum memoization file size above which parallelization is disabled. 
-	 * This is necessary to avoid creating multiple readers that are too large (e.g. sometimes 
-	 * a memoization file can be over 1GB...)
-	 */
-	private static long MAX_PARALLELIZATION_MEMO_SIZE = 1024L * 1024L * 16L;
-	
-	/**
-	 * The original URI requested for this server.
-	 */
-	private URI uri;
-	
+
 	/**
 	 * Minimum tile size - smaller values will be ignored.
 	 */
-	private static int MIN_TILE_SIZE = 32;
+	private static final int MIN_TILE_SIZE = 32;
 
 	/**
 	 * Default tile size - when no other value is available.
 	 */
-	private static int DEFAULT_TILE_SIZE = 512;
-	private static final Pattern ZARR_FILE_PATTERN = Pattern.compile("\\.zarr/?(\\d+/?)?$");
+	private static final int DEFAULT_TILE_SIZE = 512;
 
-//	/**
-//	 * Maximum tile size - larger values will be ignored.
-//	 */
-//	private static int MAX_TILE_SIZE = 4096;
-	
+	/**
+	 * The original URI requested for this server.
+	 */
+	private final URI uri;
+
 	/**
 	 * Image names (in lower case) normally associated with 'extra' images, but probably not representing the main image in the file.
 	 */
-	private static Collection<String> extraImageNames = new HashSet<>(
-			Arrays.asList("overview", "label", "thumbnail", "macro", "macro image", "macro mask image", "label image", "overview image", "thumbnail image"));
+	private static final Collection<String> extraImageNames = Set.of(
+			"overview", "label", "thumbnail", "macro", "macro image", "macro mask image", "label image", "overview image", "thumbnail image"
+	);
 	
 	/**
 	 * Original metadata, populated when reading the file.
 	 */
-	private ImageServerMetadata originalMetadata;
+	private final ImageServerMetadata originalMetadata;
 	
 	/**
 	 * Arguments passed to constructor.
 	 */
-	private String[] args;
+	private final String[] args;
 	
 	/**
 	 * File path if possible, or a URI otherwise.
@@ -171,19 +131,14 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 	private String filePathOrUrl;
 	
 	/**
-	 * Fix issue related to VSI images having (wrong) z-slices
-	 */
-//	private boolean doChannelZCorrectionVSI = false;
-	
-	/**
 	 * A map linking an identifier (image name) to series number for 'full' images.
 	 */
-	private Map<String, ServerBuilder<BufferedImage>> imageMap = null;
+	private final Map<String, ServerBuilder<BufferedImage>> imageMap;
 	
 	/**
 	 * A map linking an identifier (image name) to series number for additional images, e.g. thumbnails or macro images.
 	 */
-	private Map<String, Integer> associatedImageMap = null;
+	private final Map<String, Integer> associatedImageMap;
 	
 	/**
 	 * Numeric identifier for the image (there might be more than one in the file)
@@ -193,17 +148,7 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 	/**
 	 * Format for the current reader.
 	 */
-	private String format;
-	
-//	/**
-//	 * QuPath-specific options for how the image server should behave, such as using parallelization or memoization.
-//	 */
-//	private BioFormatsServerOptions options;
-		
-//	/**
-//	 * Manager to help keep multithreading under control.
-//	 */
-//	private static BioFormatsReaderManager manager = new BioFormatsReaderManager();
+	private final String format;
 	
 	/**
 	 * ColorModel to use with all BufferedImage requests.
@@ -213,27 +158,17 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 	/**
 	 * Pool of readers for use with this server.
 	 */
-	private ReaderPool readerPool;
-	
-	/**
-	 * Primary metadata store.
-	 */
-//	private OMEPyramidStore meta;
+	private final ReaderPool readerPool;
 	
 	/**
 	 * Cached path
 	 */
-	private String path;
+	private final String path;
 	
 	/**
 	 * Wrapper to the args passed to the reader, after parsing.
 	 */
-	private BioFormatsArgs bfArgs;
-	
-//	/**
-//	 * Try to parallelize multichannel requests (experimental!)
-//	 */
-//	private boolean parallelizeMultichannel = true;
+	private final BioFormatsArgs bfArgs;
 
 	
 	/**
@@ -259,11 +194,7 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 	 * @param options
 	 * @param args
 	 * @return
-	 * @throws FormatException
 	 * @throws IOException
-	 * @throws DependencyException
-	 * @throws ServiceException
-	 * @throws URISyntaxException
 	 */
 	static BioFormatsImageServer checkSupport(URI uri, final BioFormatsServerOptions options, String...args) throws IOException {
 		try {
@@ -272,6 +203,7 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 			// This is expected to throw an exception if it fails; without this check, Bio-Formats can appear to work
 			// but then fail when trying to read the first tile.
 			// The risk is that this will be slow for large, non-tiled images.
+			// TODO: Avoid reader access outside the reader pool
 			var reader = server.readerPool.getMainReader();
 			if (reader.getSizeX() > 0 && reader.getSizeY() > 0 && reader.openBytes(0, 0, 0, 1, 1) != null) {
 				return server;
@@ -283,13 +215,11 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 		}
 	}
 	
-	BioFormatsImageServer(URI uri, final BioFormatsServerOptions options, String...args) throws FormatException, IOException, DependencyException, ServiceException, URISyntaxException {
+	private BioFormatsImageServer(URI uri, final BioFormatsServerOptions options, String...args) throws FormatException, IOException, DependencyException, ServiceException, URISyntaxException {
 		super();
 
 		long startTime = System.currentTimeMillis();
 
-//		this.options = options;
-		
 		// Create variables for metadata
 		int width = 0, height = 0, nChannels = 1, nZSlices = 1, nTimepoints = 1, tileWidth = 0, tileHeight = 0;
 		double pixelWidth = Double.NaN, pixelHeight = Double.NaN, zSpacing = Double.NaN, magnification = Double.NaN;
@@ -348,7 +278,7 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 		List<ImageChannel> channels = new ArrayList<>();
 
 		// Create a reader & extract the metadata
-		readerPool = new ReaderPool(options, filePathOrUrl, bfArgs, channels);
+		readerPool = new ReaderPool(options, filePathOrUrl, bfArgs);
 		IFormatReader reader = readerPool.getMainReader();
 		var meta = (OMEPyramidStore)reader.getMetadataStore();
 
@@ -840,7 +770,7 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 		ImageIO.setUseCache(false);
 
 		long endTime = System.currentTimeMillis();
-		logger.debug(String.format("Initialization time: %d ms", endTime-startTime));
+		logger.debug("Initialization time: {} ms", endTime - startTime);
 	}
 	
 	
@@ -850,7 +780,7 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 	 * @param imageLength corresponding image width or height
 	 * @return a sensible tile length, bounded by the image width or height
 	 */
-	static int getDefaultTileLength(int tileLength, int imageLength) {
+	private static int getDefaultTileLength(int tileLength, int imageLength) {
 		if (tileLength <= 0) {
 			tileLength = DEFAULT_TILE_SIZE;
 		} else if (tileLength < MIN_TILE_SIZE) {
@@ -917,19 +847,6 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 		return DefaultImageServerBuilder.createInstance(BioFormatsServerBuilder.class, getMetadata(), uri, args);
 	}
 	
-	int getPreferredTileWidth() {
-		return getMetadata().getPreferredTileWidth();
-	}
-
-	int getPreferredTileHeight() {
-		return getMetadata().getPreferredTileHeight();
-	}
-
-	
-//	IFormatReader getPrimaryReader() throws DependencyException, ServiceException, FormatException, IOException {
-//		return manager.getPrimaryReader(this, this.filePath);
-//	}
-	
 	/**
 	 * Get the series index, as used by Bio-Formats.
 	 * @return
@@ -969,17 +886,16 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 	 * @return
 	 */
 	public OMEPyramidStore getMetadataStore() {
-		return readerPool.metadata;
+		return readerPool.getMetadata();
 	}
 	
 	/**
 	 * Retrieve a string representation of the metadata OME-XML.
-	 * 
 	 * @return
 	 */
 	public String dumpMetadata() {
 		try {
-			OMEXMLMetadata metadata = (OMEXMLMetadata)getMetadataStore();
+			OMEXMLMetadata metadata = getMetadataStore();
 			return metadata.dumpXML();
 		} catch (Exception e) {
 			logger.error("Unable to dump metadata", e);
@@ -1004,7 +920,7 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 		try {
 			return readerPool.openSeries(series);
 		} catch (Exception e) {
-			logger.error("Error reading associated image " + name + ": " + e.getLocalizedMessage(), e);
+            logger.error("Error reading associated image {}: {}", name, e.getLocalizedMessage(), e);
 			return null;
 		}
 	}
@@ -1032,7 +948,8 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 
 	@Override
 	public Collection<PathObject> readPathObjects() {
-		IFormatReader reader = readerPool.nextQueuedReader();
+		// TODO: Avoid reader access outside the reader pool
+		IFormatReader reader = readerPool.getMainReader();
 		if (reader == null) {
 			logger.debug("Cannot get reader. Returning no path objects");
 			return List.of();
@@ -1103,829 +1020,6 @@ public class BioFormatsImageServer extends AbstractTileableImageServer implement
 				})
 				.toList();
 	}
-	
-	/**
-	 * Get the class name of the first reader that potentially supports the file type, or null if no reader can be found.
-	 * <p>
-	 * This method only uses the path and file extensions, generously returning the first potential 
-	 * reader based on the extension. Its purpose is to help filter out hopeless cases, not to establish 
-	 * the 'correct' reader.
-	 * 
-	 * @param path
-	 * @return
-	 * @throws InstantiationException
-	 * @throws IllegalAccessException
-	 * @throws IllegalArgumentException
-	 * @throws InvocationTargetException
-	 * @throws NoSuchMethodException
-	 * @throws SecurityException
-	 */
-	static String getSupportedReaderClass(String path) throws InstantiationException, IllegalAccessException, IllegalArgumentException, InvocationTargetException, NoSuchMethodException, SecurityException {
-		path = path.toLowerCase();
-		for (var cls : ReaderPool.getDefaultClassList().getClasses()) {
-			var reader = cls.getConstructor().newInstance();
-			if (reader.isThisType(path, false))
-				return cls.getName();
-			for (String s : reader.getSuffixes()) {
-				if (s != null && !s.isBlank() && path.endsWith(s.toLowerCase()))
-					return cls.getName();
-			}
-		}
-		return null;
-	}
-
-	/**
-	 * Helper class that manages a pool of readers.
-	 * The purpose is to allow multiple threads to take the next available reader, without
-	 */
-	static class ReaderPool implements AutoCloseable {
-		
-		private static final Logger logger = LoggerFactory.getLogger(ReaderPool.class);
-
-		private static final int DEFAULT_TIMEOUT_SECONDS = 60;
-
-		/**
-		 * Absolute maximum number of permitted readers (queue capacity)
-		 */
-		private static final int MAX_QUEUE_CAPACITY = 128;
-
-		private static ClassList<IFormatReader> defaultClassList;
-		
-		private String id;
-		private BioFormatsServerOptions options;
-		private BioFormatsArgs args;
-		private ClassList<IFormatReader> classList;
-		
-		private volatile boolean isClosed = false;
-		
-		private AtomicInteger totalReaders = new AtomicInteger(0);
-		private List<IFormatReader> additionalReaders = Collections.synchronizedList(new ArrayList<>());
-		private ArrayBlockingQueue<IFormatReader> queue;
-		
-		private OMEPyramidStore metadata;
-		private IFormatReader mainReader;
-		
-		private ForkJoinTask<?> task;
-		private final List<ImageChannel> channels;
-
-		private int timeoutSeconds;
-
-		// This may be reused by OMERO extension? Not sure, but need to change cautiously...
-		ReaderPool(BioFormatsServerOptions options, String id, BioFormatsArgs args, List<ImageChannel> channels) throws FormatException, IOException {
-			this.id = id;
-			this.options = options;
-			this.args = args;
-			this.channels = channels;
-			
-			queue = new ArrayBlockingQueue<>(MAX_QUEUE_CAPACITY); // Set a reasonably large capacity (don't want to block when trying to add)
-			metadata = (OMEPyramidStore)MetadataTools.createOMEXMLMetadata();
-
-			timeoutSeconds = getTimeoutSeconds();
-			
-			// Create the main reader
-			long startTime = System.currentTimeMillis();
-			mainReader = createReader(options, null, id, metadata, args);
-			
-			long endTime = System.currentTimeMillis();
-			logger.debug("Reader {} created in {} ms", mainReader, endTime - startTime);
-			
-			// Make the main reader available
-			queue.add(mainReader);
-			
-			// Store the class so we don't need to go hunting later
-			classList = unwrapClasslist(mainReader);
-		}
-
-		/**
-		 * Make the timeout adjustable.
-		 * See https://github.com/qupath/qupath/issues/1265
-		 * @return
-		 */
-		private int getTimeoutSeconds() {
-			String timeoutString = System.getProperty("bioformats.readerpool.timeout", null);
-			if (timeoutString != null) {
-				try {
-					return Integer.parseInt(timeoutString);
-				} catch (NumberFormatException e) {
-					logger.warn("Unable to parse timeout value: {}", timeoutString, e);
-				}
-			}
-			return DEFAULT_TIMEOUT_SECONDS;
-		}
-
-		/**
-		 * Required for Bio-Formats 7.0.1 - see https://github.com/ome/bioformats/issues/4109
-		 * @return
-		 */
-		private static ClassList<IFormatReader> getDefaultClassList() {
-			if (defaultClassList == null) {
-				synchronized (BioFormatsImageServer.class) {
-					if (defaultClassList == null) {
-						ClassList<IFormatReader> classes = new ClassList<>(IFormatReader.class);
-						for (var cls : ImageReader.getDefaultReaderClasses().getClasses()) {
-							try {
-								var instance = cls.getConstructor().newInstance();
-								if (instance != null)
-									classes.addClass(cls);
-							} catch (Throwable t) {
-								t.printStackTrace();
-								logger.debug("Cannot instantiate reader class " + cls.getName() + ": " + t.getMessage());
-							}
-						}
-						defaultClassList = classes;
-					}
-				}
-			}
-			return defaultClassList;
-		}
-
-		IFormatReader getMainReader() {
-			return mainReader;
-		}
-		
-		private void createAdditionalReader(BioFormatsServerOptions options, final ClassList<IFormatReader> classList, 
-				final String id, BioFormatsArgs args) {
-			try {
-				if (isClosed)
-					return;
-				logger.debug("Requesting new reader for thread {}", Thread.currentThread());
-				var newReader = createReader(options, classList, id, null, args);
-				if (newReader != null) {
-					additionalReaders.add(newReader);
-					queue.add(newReader);
-					logger.debug("Created new reader (total={})", additionalReaders.size());
-				} else
-					logger.warn("New Bio-Formats reader could not be created (returned null)");
-			} catch (Exception e) {
-				logger.error("Error creating additional readers: " + e.getLocalizedMessage(), e);
-			}
-		}
-		
-		
-		private int getMaxReaders() {
-			int max = options == null ? Runtime.getRuntime().availableProcessors() : options.getMaxReaders();
-			return Math.min(MAX_QUEUE_CAPACITY, Math.max(1, max));
-		}
-		
-
-		/**
-		 * Create a new {@code IFormatReader}, with memoization if necessary.
-		 * 
-		 * @param options 	options used to control the reader generation
-		 * @param classList optionally specify a list of potential reader classes, if known (to avoid a more lengthy search)
-		 * @param id 		file path for the image.
-		 * @param store 	optional MetadataStore; this will be set in the reader if needed. If it is unspecified, a dummy store will be created a minimal metadata requested.
-		 * @param args      optional args to customize reading
-		 * @return the {@code IFormatReader}
-		 * @throws FormatException
-		 * @throws IOException
-		 */
-		@SuppressWarnings("resource")
-		private IFormatReader createReader(final BioFormatsServerOptions options, final ClassList<IFormatReader> classList, 
-				final String id, final MetadataStore store, BioFormatsArgs args) throws FormatException, IOException {
-			
-			int maxReaders = getMaxReaders();
-			int nReaders = totalReaders.getAndIncrement();
-			if (mainReader != null && nReaders > maxReaders) {
-				logger.warn("No new reader will be created (already created {}, max readers {})", nReaders, maxReaders);
-				totalReaders.decrementAndGet();
-				return null;
-			}
-			
-			IFormatReader imageReader;
-			Matcher zarrMatcher = ZARR_FILE_PATTERN.matcher(id.toLowerCase());
-			if (new File(id).isDirectory() || zarrMatcher.find()) {
-				// Using new ImageReader() on a directory won't work
-				imageReader = new ZarrReader();
-				if (id.startsWith("https") && imageReader.getMetadataOptions() instanceof DynamicMetadataOptions zarrOptions) {
-					zarrOptions.set("omezarr.alt_store", id);
-				}
-			} else {
-				if (classList != null) {
-					imageReader = new ImageReader(classList);
-				} else {
-					imageReader = new ImageReader(getDefaultClassList());
-				}
-			}
-
-			imageReader.setFlattenedResolutions(false);
-			
-			// Try to set any reader options that we have
-			MetadataOptions metadataOptions = imageReader.getMetadataOptions();
-			var readerOptions = args.readerOptions;
-			if (!readerOptions.isEmpty() && metadataOptions instanceof DynamicMetadataOptions) {
-				for (var option : readerOptions.entrySet()) {
-					((DynamicMetadataOptions)metadataOptions).set(option.getKey(), option.getValue());
-				}
-			}
-			
-			// TODO: Warning! Memoization does not play nicely with options like 
-			// --bfOptions zeissczi.autostitch=false
-			// in a way that options don't have an effect unless QuPath is restarted.
-			Memoizer memoizer = null;
-			int memoizationTimeMillis = options.getMemoizationTimeMillis();
-			File dir = null;
-			File fileMemo = null;
-			boolean useTempMemoDirectory = false;
-			// Check if we want to (and can) use memoization
-			if (BioFormatsServerOptions.allowMemoization() && memoizationTimeMillis >= 0) {
-				// Try to use a specified directory
-				String pathMemoization = options.getPathMemoization();
-				if (pathMemoization != null && !pathMemoization.trim().isEmpty()) {
-					dir = new File(pathMemoization);
-					if (!dir.isDirectory()) {
-						logger.warn("Memoization path does not refer to a valid directory, will be ignored: {}", dir.getAbsolutePath());
-						dir = null;
-					}
-				}
-				if (dir == null) {
-					dir = getTempMemoDir(true);
-					useTempMemoDirectory = dir != null;
-				}
-				if (dir != null) {
-					try {
-						memoizer = new Memoizer(imageReader, memoizationTimeMillis, dir);
-						fileMemo = memoizer.getMemoFile(id);
-						// The call to .toPath() should throw an InvalidPathException if there are illegal characters
-						// If so, we want to know that now before committing to the memoizer
-						if (fileMemo != null && fileMemo.toPath() != null)
-							imageReader = memoizer;
-					} catch (Exception e) {
-						logger.warn("Unable to use memoization: {}", e.getLocalizedMessage());
-						logger.debug(e.getLocalizedMessage(), e);
-						fileMemo = null;
-						memoizer = null;
-					}
-				}
-			}
-			
-			
-			if (store != null)
-				imageReader.setMetadataStore(store);
-			else {
-				imageReader.setMetadataStore(new DummyMetadata());
-				imageReader.setOriginalMetadataPopulated(false);
-			}
-			
-			var swapDimensions = args.getSwapDimensions();
-			if (swapDimensions != null)
-				logger.debug("Creating DimensionSwapper for {}", swapDimensions);
-			
-			
-			if (id != null) {
-				if (fileMemo != null) {
-					// If we're using a temporary directory, delete the memo file when app closes
-					if (useTempMemoDirectory)
-						tempMemoFiles.add(fileMemo);
-					
-					long memoizationFileSize = fileMemo == null ? 0L : fileMemo.length();
-					boolean memoFileExists = fileMemo != null && fileMemo.exists();
-					try {
-						if (swapDimensions != null)
-							imageReader = DimensionSwapper.makeDimensionSwapper(imageReader);
-						imageReader.setId(id);
-					} catch (Exception e) {
-						if (memoFileExists) {
-							logger.warn("Problem with memoization file {} ({}), will try to delete it", fileMemo.getName(), e.getLocalizedMessage());
-							fileMemo.delete();
-						}
-						imageReader.close();
-						if (swapDimensions != null)
-							imageReader = DimensionSwapper.makeDimensionSwapper(imageReader);
-						imageReader.setId(id);
-					}
-					memoizationFileSize = fileMemo == null ? 0L : fileMemo.length();
-					if (memoizationFileSize > 0L) {
-						if (memoizationFileSize > MAX_PARALLELIZATION_MEMO_SIZE) {
-							logger.warn(String.format("The memoization file is very large (%.1f MB) - parallelization may be turned off to save memory",
-									memoizationFileSize/(1024.0*1024.0)));
-						}
-						memoizationSizeMap.put(id, memoizationFileSize);
-					} if (memoizationFileSize == 0L)
-						logger.debug("No memoization file generated for {}", id);
-					else if (!memoFileExists)
-						logger.debug(String.format("Generating memoization file %s (%.2f MB)", fileMemo.getAbsolutePath(), memoizationFileSize/1024.0/1024.0));
-					else
-						logger.debug("Memoization file exists at {}", fileMemo.getAbsolutePath());
-				} else {
-					if (swapDimensions != null)
-						imageReader = DimensionSwapper.makeDimensionSwapper(imageReader);
-					imageReader.setId(id);
-				}
-			}
-						
-			if (swapDimensions != null) {
-				// The series needs to be set before swapping dimensions
-				if (args.series >= 0)
-					imageReader.setSeries(args.series);
-				((DimensionSwapper)imageReader).swapDimensions(swapDimensions);
-			}
-
-			if (isClosed) {
-				imageReader.close(false);
-				return null;
-			} else {
-				cleanables.add(cleaner.register(this,
-						new ReaderCleaner(Integer.toString(cleanables.size()+1), imageReader)));
-			}
-			
-			return imageReader;
-		}
-		
-				
-		
-		private IFormatReader nextQueuedReader() {
-			var nextReader = queue.poll();
-			if (nextReader != null) {
-				if (nextReader.getCurrentFile() != null)
-					return nextReader;
-				else
-					totalReaders.decrementAndGet();
-			}
-			synchronized (this) {
-				if (!isClosed && (task == null || task.isDone()) && totalReaders.get() < getMaxReaders()) {
-					logger.debug("Requesting reader for {}", id);
-					task = ForkJoinPool.commonPool().submit(() -> createAdditionalReader(options, classList, id, args));				
-				}
-			}
-			if (isClosed)
-				return null;
-			try {
-				var reader = queue.poll(timeoutSeconds, TimeUnit.SECONDS);
-				// See https://github.com/qupath/qupath/issues/1265
-				if (reader == null) {
-					logger.warn("Bio-Formats reader request timed out after {} seconds - returning main reader", timeoutSeconds);
-					return mainReader;
-				} else
-					return reader;
-			} catch (InterruptedException e) {
-				logger.warn("Interrupted exception when awaiting next queued reader: {}", e.getLocalizedMessage());
-				return isClosed ? null : mainReader;
-			}
-		}
-
-		private void ensureOpen(IFormatReader reader) throws IOException, FormatException {
-			if (!id.equals(reader.getCurrentFile())) {
-				reader.close();
-				reader.setFlattenedResolutions(false);
-				reader.setId(id);
-			}
-		}
-		
-		
-		BufferedImage openImage(TileRequest tileRequest, int series, int nChannels, boolean isRGB, ColorModel colorModel) throws IOException, InterruptedException {
-			int level = tileRequest.getLevel();
-			int tileX = tileRequest.getTileX();
-			int tileY = tileRequest.getTileY();
-			int tileWidth = tileRequest.getTileWidth();
-			int tileHeight = tileRequest.getTileHeight();
-			int z = tileRequest.getZ();
-			int t = tileRequest.getT();
-	
-			byte[][] bytes = null;
-			int effectiveC;
-			ByteOrder order = ByteOrder.BIG_ENDIAN;
-			boolean interleaved;
-			int pixelType;
-			boolean normalizeFloats = false;
-			int[] samplesPerPixel;
-
-			
-			IFormatReader ipReader = null;
-			try {
-				ipReader = nextQueuedReader();
-				if (ipReader == null) {
-					throw new IOException("Reader is null - was the image already closed? " + id);
-				}
-
-				// Check if this is non-zero
-				if (tileWidth <= 0 || tileHeight <= 0) {
-					throw new IOException("Unable to request pixels for region with downsampled size " + tileWidth + " x " + tileHeight);
-				}
-
-				synchronized (ipReader) {
-					ensureOpen(ipReader);
-					ipReader.setSeries(series);
-
-					// Some files provide z scaling (the number of z stacks decreases when the resolution becomes
-					// lower, like the width and height), so z needs to be updated for levels > 0
-					if (level > 0 && z > 0) {
-						ipReader.setResolution(0);
-						int zStacksFullResolution = ipReader.getSizeZ();
-						ipReader.setResolution(level);
-						int zStacksCurrentResolution = ipReader.getSizeZ();
-
-						if (zStacksFullResolution != zStacksCurrentResolution) {
-							z = (int) (z * zStacksCurrentResolution / (float) zStacksFullResolution);
-						}
 
 
-					} else {
-						ipReader.setResolution(level);
-					}
-
-					order = ipReader.isLittleEndian() ? ByteOrder.LITTLE_ENDIAN : ByteOrder.BIG_ENDIAN;
-					interleaved = ipReader.isInterleaved();
-					pixelType = ipReader.getPixelType();
-					normalizeFloats = ipReader.isNormalized();
-					samplesPerPixel = IntStream.range(0, metadata.getChannelCount(series))
-							.map(channel -> metadata.getChannelSamplesPerPixel(series, channel).getValue())
-							.toArray();
-
-					// Single-channel & RGB images are straightforward... nothing more to do
-					if ((ipReader.isRGB() && isRGB) || nChannels == 1) {
-						// Read the image - or at least the first channel
-						int ind = ipReader.getIndex(z, 0, t);
-						try {
-							byte[] bytesSimple = ipReader.openBytes(ind, tileX, tileY, tileWidth, tileHeight);
-							return AWTImageTools.openImage(bytesSimple, ipReader, tileWidth, tileHeight);
-						} catch (Exception | UnsatisfiedLinkError e) {
-							logger.warn("Unable to open image {} for {}", ind, tileRequest.getRegionRequest());
-							throw convertToIOException(e);
-						}
-					}
-					// Read bytes for all the required channels
-					effectiveC = ipReader.getEffectiveSizeC();
-					bytes = new byte[effectiveC][];
-					try {
-						for (int c = 0; c < effectiveC; c++) {
-							int ind = ipReader.getIndex(z, c, t);
-							bytes[c] = ipReader.openBytes(ind, tileX, tileY, tileWidth, tileHeight);
-						}
-					} catch (ClosedChannelException e) {
-						// This occurs when a request is interrupted
-						logger.warn("Closed channel exception, closing reader");
-						ipReader.close(false);
-						throw e;
-					} catch (Exception | UnsatisfiedLinkError e) {
-						throw convertToIOException(e);
-					}
-				}
-			} catch (FormatException e) {
-				logger.debug("Unable to open reader: {}", e.getMessage(), e);
-				throw new IOException(e);
-			} finally {
-				if (Thread.interrupted()) {
-					logger.debug("Thread interrupted, flag will be reset: {}", Thread.currentThread());
-				}
-				if (ipReader != null)
-					queue.put(ipReader);
-			}
-
-			OMEPixelParser omePixelParser = new OMEPixelParser.Builder()
-					.isInterleaved(interleaved)
-					.pixelType(switch (pixelType) {
-						case FormatTools.UINT8 -> PixelType.UINT8;
-						case FormatTools.INT8 -> PixelType.INT8;
-						case FormatTools.UINT16 -> PixelType.UINT16;
-						case FormatTools.INT16 -> PixelType.INT16;
-						case FormatTools.UINT32 -> PixelType.UINT32;
-						case FormatTools.INT32 -> PixelType.INT32;
-						case FormatTools.FLOAT -> PixelType.FLOAT32;
-						case FormatTools.DOUBLE -> PixelType.FLOAT64;
-						default -> throw new IllegalStateException("Unexpected value: " + pixelType);
-					})
-					.byteOrder(order)
-					.normalizeFloats(normalizeFloats)
-					.effectiveNChannels(effectiveC)
-					.samplesPerPixel(samplesPerPixel)
-					.build();
-			
-			return omePixelParser.parse(bytes, tileWidth, tileHeight, nChannels, colorModel);
-		}
-
-		/**
-		 * Ensure a throwable is an IOException.
-		 * This gives the opportunity to include more human-readable messages for common errors.
-		 * @param t
-		 * @return
-		 */
-		private static IOException convertToIOException(Throwable t) {
-			if (GeneralTools.isMac()) {
-				String message = t.getMessage();
-				if (message != null) {
-					if (message.contains("ome.jxrlib.JXRJNI")) {
-						return new IOException("Bio-Formats does not support JPEG-XR on Apple Silicon: " + t.getMessage(), t);
-					}
-					if (message.contains("org.libjpegturbo.turbojpeg.TJDecompressor")) {
-						return new IOException("Bio-Formats does not currently support libjpeg-turbo on Apple Silicon", t);
-					}
-				}
-			}
-			if (t instanceof IOException e)
-				return e;
-			return new IOException(t);
-		}
-		
-		
-		public BufferedImage openSeries(int series) throws InterruptedException, FormatException, IOException {
-			IFormatReader reader = null;
-			try {
-				reader = nextQueuedReader();
-				synchronized (reader) {
-					int previousSeries = reader.getSeries();
-					try {
-						reader.setSeries(series);
-						int nResolutions = reader.getResolutionCount();
-						if (nResolutions > 0) {
-							reader.setResolution(0);
-						}
-						// TODO: Handle color transforms here, or in the display of labels/macro images - in case this isn't RGB
-						byte[] bytesSimple = reader.openBytes(reader.getIndex(0, 0, 0));
-						return AWTImageTools.openImage(bytesSimple, reader, reader.getSizeX(), reader.getSizeY());
-					} finally {
-						reader.setSeries(previousSeries);
-					}
-				}
-			} finally {
-				queue.put(reader);
-			}
-		}
-		
-		
-		
-		private static ClassList<IFormatReader> unwrapClasslist(IFormatReader reader) {
-			while (reader instanceof ReaderWrapper || reader instanceof ImageReader) {
-				if (reader instanceof ReaderWrapper wrapper)
-					reader = wrapper.getReader();
-				else
-					reader = ((ImageReader)reader).getReader();
-			}
-			var classlist = new ClassList<>(IFormatReader.class);
-			classlist.addClass(reader.getClass());
-			return classlist;
-		}
-		
-		
-		
-		@Override
-		public void close() throws Exception {
-			logger.debug("Closing ReaderManager");
-			isClosed = true;
-			if (task != null && !task.isDone())
-				task.cancel(true);
-			for (var c : cleanables) {
-				try {
-					c.clean();
-				} catch (Exception e) {
-					logger.error("Exception during cleanup: {}", e.getMessage(), e);
-				}
-			}
-		}
-
-		
-		private static final Cleaner cleaner = Cleaner.create();
-		private final List<Cleanable> cleanables = new ArrayList<>();
-
-
-		/**
-		 * Map of memoization file sizes.
-		 */
-		private static Map<String, Long> memoizationSizeMap = new ConcurrentHashMap<>();
-
-		/**
-		 * Temporary directory for storing memoization files
-		 */
-		private static File dirMemoTemp = null;
-
-		/**
-		 * Set of created temp memo files
-		 */
-		private static Set<File> tempMemoFiles = new HashSet<>();
-
-
-		/**
-		 * Request the file size of any known memoization file for a specific ID.
-		 * 
-		 * @param id
-		 * @return
-		 */
-		public long getMemoizationFileSize(String id) {
-			return memoizationSizeMap.getOrDefault(id, Long.valueOf(0L));
-		}
-
-		private static File getTempMemoDir(boolean create) throws IOException {
-			if (create && dirMemoTemp == null) {
-				synchronized (ReaderPool.class) {
-					if (dirMemoTemp == null) {
-						Path path = Files.createTempDirectory("qupath-memo-");
-						dirMemoTemp = path.toFile();
-						Runtime.getRuntime().addShutdownHook(new Thread() {
-							@Override
-							public void run() {
-								deleteTempMemoFiles();
-							}
-						});
-						logger.warn("Temp memoization directory created at {}", dirMemoTemp);
-						logger.warn("If you want to avoid this warning, either specify a memoization directory in the preferences or turn off memoization by setting the time to < 0");
-					}
-				}
-			}
-			return dirMemoTemp;
-		}
-
-		/**
-		 * Delete any memoization files registered as being temporary, and also the 
-		 * temporary memoization directory (if it exists).
-		 * Note that this acts both recursively and rather conservatively, stopping if a file is 
-		 * encountered that is not expected.
-		 */
-		private static void deleteTempMemoFiles() {
-			for (File f : tempMemoFiles) {
-				// Be extra-careful not to delete too much...
-				if (!f.exists())
-					continue;
-				if (!f.isFile() || !f.getName().endsWith(".bfmemo")) {
-					logger.warn("Unexpected memoization file, will not delete {}", f.getAbsolutePath());
-					return;
-				}
-				if (f.delete())
-					logger.debug("Deleted temp memoization file {}", f.getAbsolutePath());
-				else
-					logger.warn("Could not delete temp memoization file {}", f.getAbsolutePath());
-			}
-			if (dirMemoTemp == null)
-				return;
-			deleteEmptyDirectories(dirMemoTemp);
-		}
-
-		/**
-		 * Delete a directory and all sub-directories, assuming each contains only empty directories.
-		 * This is applied recursively, stopping at the first failure (i.e. any directory containing files).
-		 * @param dir
-		 * @return true if the directory could be deleted, false otherwise
-		 */
-		private static boolean deleteEmptyDirectories(File dir) {
-			if (!dir.isDirectory())
-				return false;
-			int nFiles = 0;
-			var files = dir.listFiles();
-			if (files == null) {
-				logger.debug("Unable to list files for {}", dir);
-				return false;
-			}
-			for (File f : files) {
-				if (f.isDirectory()) {
-					if (!deleteEmptyDirectories(f))
-						return false;
-				} else if (f.isFile())
-					nFiles++;
-			}
-			if (nFiles == 0) {
-				if (dir.delete()) {
-					logger.debug("Deleting empty memoization directory {}", dir.getAbsolutePath());
-					return true;
-				} else {
-					logger.warn("Could not delete temp memoization directory {}", dir.getAbsolutePath());
-					return false;
-				}
-			} else {
-				logger.warn("Temp memoization directory contains files, will not delete {}", dir.getAbsolutePath());
-				return false;
-			}
-		}
-
-
-		/**
-		 * Helper class that helps ensure readers are closed when a reader pool is no longer reachable.
-		 */
-		static class ReaderCleaner implements Runnable {
-
-			private final String name;
-			private final IFormatReader reader;
-
-			ReaderCleaner(String name, IFormatReader reader) {
-				this.name = name;
-				this.reader = reader;
-			}
-
-			@Override
-			public void run() {
-                logger.debug("Cleaner {} called for {} ({})", name, reader, reader.getCurrentFile());
-				try {
-					this.reader.close(false);
-				} catch (IOException e) {
-                    logger.warn("Error when calling cleaner for {}", name, e);
-				}
-			}
-
-		}
-		
-		
-	}
-	
-	
-	
-	
-	
-	
-	static class BioFormatsArgs {
-		
-		@Option(names = {"--series", "-s"}, defaultValue = "-1", description = "Series number (0-based, must be < image count for the file)")
-		int series = -1;
-
-		@Option(names = {"--name", "-n"}, defaultValue = "", description = "Series name (legacy option, please use --series instead)")
-		String seriesName = "";
-		
-		@Option(names = {"--dims"}, defaultValue = "", description = "Swap dimensions. "
-				+ "This should be a String of the form XYCZT, ordered according to how the image plans should be interpreted.")
-		String swapDimensions = null;
-
-		// Specific options used by some Bio-Formats readers, e.g. Map.of("zeissczi.autostitch", "false")
-		@Option(names = {"--bfOptions"}, description = "Bio-Formats reader options")
-		Map<String, String> readerOptions = new LinkedHashMap<>();
-		
-		@Unmatched
-		List<String> unmatched = new ArrayList<>();
-		
-		BioFormatsArgs() {}
-		
-		/**
-		 * Return to an array of String args.
-		 * @param series 
-		 * @return
-		 */
-		String[] backToArgs(int series) {
-			var args = new ArrayList<String>();
-			if (series >= 0) {
-				args.add("--series");
-				args.add(Integer.toString(series));
-			} else if (this.series >= 0) {
-				args.add("--series");
-				args.add(Integer.toString(this.series));				
-			} else if (seriesName != null && !seriesName.isBlank()) {
-				args.add("--name");
-				args.add(seriesName);				
-			}
-			if (swapDimensions != null && !swapDimensions.isBlank()) {
-				args.add("--dims");
-				args.add(swapDimensions);
-			}
-			for (var option : readerOptions.entrySet()) {
-				// Note: this assumes that options & values contain no awkwardness (e.g. quotes, spaces)
-				args.add("--bfOptions");
-				args.add(option.getKey()+"="+option.getValue());
-			}
-			args.addAll(unmatched);
-			return args.toArray(String[]::new);
-		}
-		
-		String getSwapDimensions() {
-			return swapDimensions == null || swapDimensions.isBlank() ? null : swapDimensions.toUpperCase();
-		}
-		
-		static BioFormatsArgs parse(String[] args) {
-			var bfArgs = new BioFormatsArgs();
-			new CommandLine(bfArgs).parseArgs(args);
-			return bfArgs;
-		}
-
-		@Override
-		public int hashCode() {
-			final int prime = 31;
-			int result = 1;
-			result = prime * result + ((readerOptions == null) ? 0 : readerOptions.hashCode());
-			result = prime * result + series;
-			result = prime * result + ((seriesName == null) ? 0 : seriesName.hashCode());
-			result = prime * result + ((swapDimensions == null) ? 0 : swapDimensions.hashCode());
-			result = prime * result + ((unmatched == null) ? 0 : unmatched.hashCode());
-			return result;
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if (this == obj)
-				return true;
-			if (obj == null)
-				return false;
-			if (getClass() != obj.getClass())
-				return false;
-			BioFormatsArgs other = (BioFormatsArgs) obj;
-			if (readerOptions == null) {
-				if (other.readerOptions != null)
-					return false;
-			} else if (!readerOptions.equals(other.readerOptions))
-				return false;
-			if (series != other.series)
-				return false;
-			if (seriesName == null) {
-				if (other.seriesName != null)
-					return false;
-			} else if (!seriesName.equals(other.seriesName))
-				return false;
-			if (swapDimensions == null) {
-				if (other.swapDimensions != null)
-					return false;
-			} else if (!swapDimensions.equals(other.swapDimensions))
-				return false;
-			if (unmatched == null) {
-				if (other.unmatched != null)
-					return false;
-			} else if (!unmatched.equals(other.unmatched))
-				return false;
-			return true;
-		}
-		
-	}
-	
 }
