+++ qupath-extension-bioformats/src/main/java/qupath/lib/images/servers/bioformats/ReaderCleaner.java
@@ -1,0 +1,33 @@
+package qupath.lib.images.servers.bioformats;
+
+import java.io.IOException;
+import loci.formats.IFormatReader;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+/**
+ * Helper class that helps ensure readers are closed when a reader pool is no longer reachable.
+ */
+class ReaderCleaner implements Runnable {
+
+    private static final Logger logger = LoggerFactory.getLogger(ReaderCleaner.class);
+
+    private final String name;
+    private final IFormatReader reader;
+
+    ReaderCleaner(String name, IFormatReader reader) {
+        this.name = name;
+        this.reader = reader;
+    }
+
+    @Override
+    public void run() {
+        logger.info("Cleaner {} called for {} ({})", name, reader, reader.getCurrentFile());
+        try {
+            this.reader.close(false);
+        } catch (IOException e) {
+            logger.warn("Error when calling cleaner for {}", name, e);
+        }
+    }
+
+}
